Fixed typos. Now an analysis with all its PDFs and PDF ratios and LLH ratio functions can be constructed.
git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/skylab/branches/skylab3@165312 16731396-06f5-0310-8873-f7f720988828

+++ skylab/core/source_hypothesis.py
@@ -163,7 +163,7 @@ class SourceHypoGroupManager(object):
             The SourceHypoGroup instance for which the map array should get
             extented.
         """
-        arr = np.empty((shg.n_sources,2))
+        arr = np.empty((shg.n_sources,2), dtype=np.int)
         arr[:,0] = self.n_src_hypo_groups-1 # Group index.
         arr[:,1] = np.arange(shg.n_sources) # Group source index.
         self._sidx_to_gidx_gsidx_map_arr = np.vstack(
